Update and fixes for 2021 August 09 (#248)
* Ensure TKey used by AddIfNotDefault(...) extension is not null

* Added class constraint to generic types for events

* Change EventContext.Event to not be nullable

* Log warning when either the envelope of or it's event is null

* Introduce DeserializationContext

* Standardize generation of sequence number in QueueStorage

* Introduced AzureQueueStorageSchedulingId to ease the parsing and generation

* Make IServiceScope argument in SerializeAsync(...) and DeserializeAsync(...) of EventBusTransportBase the first

* Remove BeginLoggingScopeForConsume(...) overload that takes longg sequenceNumber

* Use offset for identifier in KafkaTransport

* Validate that the stream passed to DeserializationContext is readable

* Remove worflow_dispatch from build and release workflow

* Added workflow for preview releases

* Few renames in workflow files

+++ src/Tingle.EventBus.Transports.Azure.QueueStorage/AzureQueueStorageTransport.cs
@@ -22,8 +22,6 @@ namespace Tingle.EventBus.Transports.Azure.QueueStorage
     [TransportName(TransportNames.AzureQueueStorage)]
     public class AzureQueueStorageTransport : EventBusTransportBase<AzureQueueStorageTransportOptions>, IDisposable
     {
-        private const string SequenceNumberSeparator = "|";
-
         private readonly Dictionary<(Type, bool), QueueClient> queueClientsCache = new();
         private readonly SemaphoreSlim queueClientsCacheLock = new(1, 1); // only one at a time.
         private readonly CancellationTokenSource stoppingCts = new();
@@ -103,10 +101,10 @@ namespace Tingle.EventBus.Transports.Azure.QueueStorage
         {
             using var scope = CreateScope();
             using var ms = new MemoryStream();
-            await SerializeAsync(body: ms,
+            await SerializeAsync(scope: scope,
+                                 body: ms,
                                  @event: @event,
                                  registration: registration,
-                                 scope: scope,
                                  cancellationToken: cancellationToken);
 
 
@@ -127,7 +125,7 @@ namespace Tingle.EventBus.Transports.Azure.QueueStorage
 
             // return the sequence number; both MessageId and PopReceipt are needed to update or delete
             return scheduled != null
-                    ? new ScheduledResult(id: string.Join(SequenceNumberSeparator, response.Value.MessageId, response.Value.PopReceipt), scheduled: scheduled.Value)
+                    ? new ScheduledResult(id: (AzureQueueStorageSchedulingId)response.Value, scheduled: scheduled.Value)
                     : null;
         }
 
@@ -148,10 +146,10 @@ namespace Tingle.EventBus.Transports.Azure.QueueStorage
             foreach (var @event in events)
             {
                 using var ms = new MemoryStream();
-                await SerializeAsync(body: ms,
+                await SerializeAsync(scope: scope,
+                                     body: ms,
                                      @event: @event,
                                      registration: registration,
-                                     scope: scope,
                                      cancellationToken: cancellationToken);
                 // if scheduled for later, calculate the visibility timeout
                 var visibilityTimeout = scheduled - DateTimeOffset.UtcNow;
@@ -167,7 +165,7 @@ namespace Tingle.EventBus.Transports.Azure.QueueStorage
                                                                   timeToLive: ttl,
                                                                   cancellationToken: cancellationToken);
                 // collect the sequence number
-                sequenceNumbers.Add(string.Join(SequenceNumberSeparator, response.Value.MessageId, response.Value.PopReceipt));
+                sequenceNumbers.Add((AzureQueueStorageSchedulingId)response.Value);
             }
 
             // return the sequence number
@@ -185,12 +183,19 @@ namespace Tingle.EventBus.Transports.Azure.QueueStorage
             }
 
             var queueClient = await GetQueueClientAsync(reg: registration, deadletter: false, cancellationToken: cancellationToken);
-            var parts = id.Split(SequenceNumberSeparator);
-            string messageId = parts[0], popReceipt = parts[1];
-            Logger.LogInformation("Cancelling '{MessageId}|{PopReceipt}' on '{QueueName}'", messageId, popReceipt, queueClient.Name);
-            await queueClient.DeleteMessageAsync(messageId: messageId,
-                                                 popReceipt: popReceipt,
-                                                 cancellationToken: cancellationToken);
+            if (AzureQueueStorageSchedulingId.TryParse(id, out var sid))
+            {
+                var messageId = sid.MessageId;
+                var popReceipt = sid.PopReceipt;
+                Logger.LogInformation("Cancelling '{MessageId}|{PopReceipt}' on '{QueueName}'", messageId, popReceipt, queueClient.Name);
+                await queueClient.DeleteMessageAsync(messageId: messageId,
+                                                     popReceipt: popReceipt,
+                                                     cancellationToken: cancellationToken);
+            }
+            else
+            {
+                Logger.LogWarning("The provided id '{Id}' does not match the expected format.", id);
+            }
         }
 
         /// <inheritdoc/>
@@ -203,22 +208,26 @@ namespace Tingle.EventBus.Transports.Azure.QueueStorage
                 throw new ArgumentNullException(nameof(ids));
             }
 
-            var splits = ids.Select(i =>
-            {
-                var parts = i.Split(SequenceNumberSeparator);
-                return (messageId: parts[0], popReceipt: parts[1]);
-            }).ToList();
-
             // log warning when doing batch
             Logger.LogWarning("Azure Queue Storage does not support batching. The events will be canceled one by one");
 
             var queueClient = await GetQueueClientAsync(reg: registration, deadletter: false, cancellationToken: cancellationToken);
-            foreach (var (messageId, popReceipt) in splits)
+
+            foreach (var id in ids)
             {
-                Logger.LogInformation("Cancelling '{MessageId}|{PopReceipt}' on '{QueueName}'", messageId, popReceipt, queueClient.Name);
-                await queueClient.DeleteMessageAsync(messageId: messageId,
-                                                     popReceipt: popReceipt,
-                                                     cancellationToken: cancellationToken);
+                if (AzureQueueStorageSchedulingId.TryParse(id, out var sid))
+                {
+                    var messageId = sid.MessageId;
+                    var popReceipt = sid.PopReceipt;
+                    Logger.LogInformation("Cancelling '{MessageId}|{PopReceipt}' on '{QueueName}'", messageId, popReceipt, queueClient.Name);
+                    await queueClient.DeleteMessageAsync(messageId: messageId,
+                                                         popReceipt: popReceipt,
+                                                         cancellationToken: cancellationToken);
+                }
+                else
+                {
+                    Logger.LogWarning("The provided id '{Id}' does not match the expected format.", id);
+                }
             }
         }
 
@@ -344,10 +353,11 @@ namespace Tingle.EventBus.Transports.Azure.QueueStorage
 
             Logger.LogDebug("Processing '{MessageId}' from '{QueueName}'", messageId, queueClient.Name);
             using var ms = new MemoryStream(Encoding.UTF8.GetBytes(message.MessageText));
-            var context = await DeserializeAsync<TEvent>(body: ms,
-                                                         contentType: null, // There is no way to get this yet
+            var context = await DeserializeAsync<TEvent>(scope: scope,
+                                                         body: ms, // There is no way to get this yet
+                                                         contentType: null,
                                                          registration: reg,
-                                                         scope: scope,
+                                                         identifier: (AzureQueueStorageSchedulingId)message,
                                                          cancellationToken: cancellationToken);
 
             Logger.LogInformation("Received message: '{MessageId}' containing Event '{Id}' from '{QueueName}'",
